Step 15 - Admin Panel/requirements.txt
Django
django-admin-rangefilter
